+++ src/impl/word_prune.c
@@ -111,9 +111,9 @@ void add_words_without_playthrough(const KWG* kwg, uint32_t node_index,
   if (node_index == 0) {
     return;
   }
-  for (int i = node_index;; i++) {
+  for (uint32_t i = node_index;; i++) {
     const uint32_t node = kwg_node(kwg, i);
-    const int ml = kwg_node_tile(node);
+    const uint8_t ml = kwg_node_tile(node);
     const int new_node_index = kwg_node_arc_index(node);
     if ((rack_get_letter(rack, ml) > 0) ||
         (rack_get_letter(rack, BLANK_MACHINE_LETTER) > 0)) {
@@ -158,7 +158,7 @@ void playthrough_words_recursive_gen(const BoardRow* board_row, const KWG* kwg,
     const uint8_t ml = current_letter;  // already unblanked
     uint32_t next_node_index = 0;
     bool accepts = false;
-    for (int i = node_index;; i++) {
+    for (uint32_t i = node_index;; i++) {
       const uint32_t node = kwg_node(kwg, i);
       if (kwg_node_tile(node) == ml) {
         next_node_index = kwg_node_arc_index(node);
@@ -174,7 +174,7 @@ void playthrough_words_recursive_gen(const BoardRow* board_row, const KWG* kwg,
                             rightstrip, leftmost_col, tiles_played, strip,
                             possible_word_list);
   } else if (!rack_is_empty(rack)) {
-    for (int i = node_index;; i++) {
+    for (uint32_t i = node_index;; i++) {
       const uint32_t node = kwg_node(kwg, i);
       const uint8_t ml = kwg_node_tile(node);
       if (ml != SEPARATION_MACHINE_LETTER) {
